// App/Containers/Discounts/DiscountAddProducts/DiscountAddProducts.type.ts
export enum DiscountProductAddTypeEnums {
  APPLIES_TO_SPECIFIC_PRODUCTS = 'APPLIES_TO_SPECIFIC_PRODUCTS',
  APPLIES_TO_SPECIFIC_CATEGORIES = 'APPLIES_TO_SPECIFIC_CATEGORIES',
  CUSTOMER_BUYS_SPECIFIC_PRODUCTS = 'CUSTOMER_BUYS_SPECIFIC_PRODUCTS',
  CUSTOMER_GETS_SPECIFIC_PRODUCTS = 'CUSTOMER_GETS_SPECIFIC_PRODUCTS',
}